move controllers into api v1; clear up unused comments
