+++ app/Http/Controllers/api/v1/NoteController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\api\v1;
 
+use App\Http\Controllers\Controller;
 use App\Models\Note;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
