+++ routes/api.php
@@ -9,9 +9,9 @@ use App\Http\Controllers\api\v1\CharacterComboController;
 use App\Http\Controllers\api\v1\CharacterMoveController;
 use App\Http\Controllers\api\v1\UserController;
 use App\Http\Controllers\Controller;
-use App\Http\Controllers\LoginController;
-use App\Http\Controllers\NoteController;
-use App\Http\Controllers\TagController;
+use App\Http\Controllers\api\v1\LoginController;
+use App\Http\Controllers\api\v1\NoteController;
+use App\Http\Controllers\api\v1\TagController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -26,17 +26,8 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-// Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-//     return $request->user();
-// });
-// Route::post('/users/')
-
-
-
-
 Route::prefix('v1')->group(function () {
     Route::post('/login', [LoginController::class, 'store']);
-
 
     // Users
     Route::get('/users', [UserController::class, 'index']);
@@ -58,7 +49,6 @@ Route::prefix('v1')->group(function () {
 
     // Moves
     Route::get('/games/{game}/characters/{character}/moves', [CharacterMoveController::class, 'index']);
-    // Route::get('/games/{game}/tags', )
 
     // Characters
     Route::get('/games/{game}/characters', [CharacterController::class, 'index']);
@@ -71,22 +61,12 @@ Route::prefix('v1')->group(function () {
 Route::middleware('auth:sanctum')->group(function (){
     Route::prefix('v1')->group(function () {
 
-        // // Moves
-        // Route::get('/games/{game}/characters/{character}/moves', [CharacterMoveController::class, 'index']);
-        // // Route::get('/games/{game}/tags', )
-
-        // // Characters
-        // Route::get('/games/{game}/characters', [CharacterController::class, 'index']);
-        // Route::get('/games/{game}/characters/{character}', [CharacterController::class, 'show']);
-
         //Combos
         Route::get('/games/{game}/characters/{character}/character-combos', [CharacterComboController::class, 'index']);
         Route::get('/games/{game}/characters/{character}/character-combos/{character_combo}', [CharacterComboController::class, 'show']);
         Route::post('/games/{game}/characters/{character}/character-combos', [CharacterComboController::class, 'store']);
         Route::put('/games/{game}/characters/{character}/character-combos/{character_combo}', [CharacterComboController::class, 'update']);
         Route::delete('/games/{game}/characters/{character}/character-combos/{character_combo}', [CharacterComboController::class, 'delete']);
-
-
 
         // Tags
         Route::get('/games/{game}/tags', [TagController::class, 'index']);
@@ -95,17 +75,11 @@ Route::middleware('auth:sanctum')->group(function (){
         Route::post('/games/{game}/characters/{character}/combos/{combo}/tags', [CharacterComboController::class, 'addCharacterComboTag']);
         Route::delete('/games/{game}/characters/{character}/combos/{combo}/tags/{tag}', [CharacterComboController::class, 'removeCharacterComboTag']);
 
-        // // Inputs
-        // Route::get('/directional-inputs', [DirectionalInputController::class, 'index']);
-        // Route::get('/games/{game}/attack-buttons', [AttackButtonController::class, 'index']);
-
         // Game Notes
         Route::get('/games/{game}/notes', [NoteController::class, 'gameNoteIndex']);
         Route::post('/games/{game}/notes', [GameController::class, 'addNote']);
         Route::put('/games/{game}/notes/{note}', [NoteController::class, 'updateGameNote']);
         Route::delete('/games/{game}/notes/{note}', [NoteController::class, 'deleteGameNote']);
-
-
 
         // Character Notes
         Route::post('/games/{game}/characters/{character}/notes', [CharacterController::class, 'addNote']);
